fix(plot): restore double precision and the tick guards in the axis path
A systematic scan of every Zig global definition against its C header
declaration (245 checked) found four size mismatches, all four in the plot
owner: graph_dx, graph_dy, tick_int_x and tick_int_y are `double` upstream
(plotstat.h:24/25/36/37, plotstat.c:14/15/23/24) and were ported as f32. The
demotion had spread through the whole axis path:

  - auto_tick was f32(f32); upstream is double(double).
  - The tick loop variables x/y and the axis bounds fx_min/fx_max/fy_min/fy_max
    stepped in f32, via a realToFloatL that upstream writes as realToDouble.
    Upstream's own comment at plotstat.c:606 says why this matters: "tick loops
    step in double; the interval underflows to 0 only beyond 1e+-308 and the
    tick_int_* > 0 guards then skip the ticks". In f32 that limit is 1e+-38.
  - Those `tick_int_x > 0` / `tick_int_y > 0` guards (plotstat.c:626/657/699/737)
    were dropped from all four arms, and upstream's `else if(tick_int_x > 0)`
    had become a bare `else`. Without them a zero interval makes `x += tick_int_x`
    a non-terminating loop -- exactly the case the guards exist to skip.
  - graphAxisDraw carried a FLoatingMin/FLoatingMax early return that upstream
    does not have; it reads as a workaround for the f32 range, and is left in
    place (now near-dead at double) rather than removed in the same change.

auto_tick was also stale: upstream replaced its "%.1e" string surgery with a
numeric mantissa/exponent split (pow/log10/floor) because the string form pulled
in _printf_float, which is no longer linked on DMCP. The current version is
ported here, at double.

The backup codec followed the same demotion: it wrote 4 bytes tagged "float"
where upstream writes sizeof(double) tagged "double" (saveRestoreBackup.c:416/
417/423/424), so a z47 backup.cfg was not interchangeable with an upstream one
for these four fields. Widened to 8 bytes and "double".

log10/floor go through the @log10/@floor builtins and pow through the sibling
plot owner's existing declaration, so extern_fn_sites stays at its ceiling.

Every pinned plot bitmap hash still matches, so this changes no rendering on the
green lanes. Whether it also settles the open Windows graphs_cov EC=28 failures
is not established -- these are real divergences from the pinned upstream either
way, and they are the first found by systematic scanning rather than inference.

zig build test 12757/12757 exit 0; all seven gates pass.

(cherry picked from commit 8d038aa379807fc0f59a94a05728528f2ae58e4b)

+++ zig_src/core/persist/calc_state_restore.zig
@@ -311,7 +311,7 @@ extern var lastIntegerBase: u32;
 extern var amortP1: u16;
 extern var amortP2: u16;
 extern var lrChosen: u16;
-extern var graph_dx: f32;
+extern var graph_dx: f64;
 extern var graph_dy: f32;
 extern var roundedTicks: u8;
 extern var PLOT_AXIS: u8;
